graphene_filtering/dev_requirements.txt
django==3.1
django-filter==21.1
factory-boy==3.2.1
graphene-django==2.15.0
pre-commit==2.17.0
black==22.3.0
isort==5.10.1
colorama==0.4.4
flake8==4.0.1